+++ internal/pkg/service/buffer/condition/checker.go
@@ -2,7 +2,6 @@ package condition
 
 import (
 	"context"
-	"fmt"
 	"sync"
 	"time"
 
@@ -170,55 +169,6 @@ func (c *Checker) check(ctx context.Context) {
 	})
 
 	c.logger.Debugf(`checked "%d" opened slices | %s`, checked, c.clock.Since(now))
-}
-
-func (c *Checker) shouldImport(ctx context.Context, now time.Time, sliceKey key.SliceKey, uploadCredExp time.Time) (ok bool, reason string, err error) {
-	// Check minimal interval
-	if interval := now.Sub(sliceKey.FileKey.OpenedAt()); interval < c.config.MinimalImportInterval {
-		reason = fmt.Sprintf(`interval "%s" is less than the MinimalImportInterval "%s"`, interval, c.config.MinimalImportInterval)
-		return false, reason, nil
-	}
-
-	// Check credentials CredExpiration
-	if uploadCredExp.Sub(now) <= MinimalCredentialsExpiration {
-		reason = fmt.Sprintf("upload credentials will expire soon, at %s", uploadCredExp.UTC().String())
-		return true, reason, nil
-	}
-
-	// Get import conditions
-	export, found := c.exports.Get(sliceKey.ExportKey.String())
-	if !found {
-		reason = "import conditions not found"
-		return false, reason, nil
-	}
-
-	// Get file stats
-	fileStats, err := c.cachedStats.FileStats(ctx, sliceKey.FileKey)
-	if err != nil {
-		return false, "", err
-	}
-
-	// Evaluate import conditions
-	ok, reason = evaluate(export.ImportConditions, now, sliceKey.FileKey.OpenedAt(), fileStats.Total)
-	return ok, reason, nil
-}
-
-func (c *Checker) shouldUpload(ctx context.Context, now time.Time, sliceKey key.SliceKey) (ok bool, reason string, err error) {
-	// Check minimal interval
-	if interval := now.Sub(sliceKey.OpenedAt()); interval < c.config.MinimalUploadInterval {
-		reason = fmt.Sprintf(`interval "%s" is less than the MinimalUploadInterval "%s"`, interval, c.config.MinimalUploadInterval)
-		return false, reason, nil
-	}
-
-	// Get slice stats
-	sliceStats, err := c.cachedStats.SliceStats(ctx, sliceKey)
-	if err != nil {
-		return false, "", err
-	}
-
-	// Evaluate upload conditions
-	ok, reason = evaluate(c.config.UploadConditions, now, sliceKey.OpenedAt(), sliceStats.Total)
-	return ok, reason, nil
 }
 
 func (c *Checker) watchExports(ctx context.Context, wg *sync.WaitGroup) (errCh <-chan error) {
